refactor(library): move launchWithModelPicker into quickActions (fix shared→features boundary)
launchWithModelPicker imported @/features/chat/tabs/tabShared and @/main,
making it a shared→features/shared→main zone violation. Moved to
features/quickActions/ where features→features is allowed. The two callers
(launchQuickAction.ts and launchLoopPrompt.ts) and the unit test import
are updated accordingly.
