+++ src/features/quickActions/launchQuickAction.ts
@@ -2,7 +2,7 @@ import { type TAbstractFile } from 'obsidian';
 
 import { t } from '@/i18n/i18n';
 import type SpecoratorPlugin from '@/main';
-import { launchWithModelPicker } from '@/shared/launchWithModelPicker';
+import { launchWithModelPicker } from './launchWithModelPicker';
 
 import { quickActionStemFromPath } from './quickActionStem';
 import { runQuickActionForFile } from './runQuickActionForFile';
